+++ example/lib/main.dart
@@ -2,6 +2,6 @@ import 'package:flutter/material.dart';
 
 import 'src/minimal_app.dart';
 
-void main() async {
+void main() {
   runApp(const MinimalApp());
 }