+++ example/lib/src/counter/notifiers/counter_notifier.dart
@@ -6,8 +6,7 @@ class CounterNotifier extends MMNotifier<int> {
   void increment() => notify(state + 1);
 }
 
-final MMManager<CounterNotifier> counterManager =
-    MMManager(CounterNotifier.new);
+final MMManager<CounterNotifier> counterManager = MMManager(CounterNotifier.new);
 
 final MMManager<CounterNotifier> disposableCounterManager =
     MMManager(CounterNotifier.new, autodispose: true);