+++ test/src/utils/test_common.dart
@@ -37,8 +37,7 @@ class TNotifier extends MMNotifier<TState> {
 
   void increment() => notify(state.copyWith(value: state.value + 1));
 
-  void append(final String suffix) =>
-      notify(state.copyWith(text: '${state.text}$suffix'));
+  void append(final String suffix) => notify(state.copyWith(text: '${state.text}$suffix'));
 
   bool get isDisposed => disposed;
 }